+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/FlowEditor.vue
@@ -1,6 +1,5 @@
 <template>
   <div class="dk-flow-editor">
-    <div>Nothing here yet.</div>
     <KButton
       appearance="secondary"
       @click="modalOpen = true"
@@ -8,98 +7,58 @@
       {{ t('plugins.free-form.datakit.flow_editor.cta') }}
     </KButton>
     <EditorModal v-model:open="modalOpen" />
+    <div class="field">
+      <BooleanField
+        name="config.debug"
+      />
+    </div>
   </div>
 </template>
 
 <script setup lang="ts">
+import { watch } from 'vue'
 import { createI18n } from '@kong-ui-public/i18n'
 import english from '../../../../locales/en.json'
 import { provideEditorStore } from '../composables'
-import type { ConfigNodeName, FieldName, NameConnection, NodeName } from '../types'
+import BooleanField from '../../shared/BooleanField.vue'
 import EditorModal from './modal/EditorModal.vue'
+
+import type { DatakitConfig } from '../types'
 
 const { t } = createI18n<typeof english>('en-us', english)
 
-defineProps<{
+const { config } = defineProps<{
   editing: boolean
-  config: any
+  config?: DatakitConfig
 }>()
 
-// todo(zehao): mock data, remove later
-// provideEditorStore(MockData.configNodes, MockData.uiNodes)
+const emit = defineEmits<{
+  change: [config: DatakitConfig]
+  error: [msg: string]
+}>()
 
-const { modalOpen } = provideEditorStore([
-  {
-    name: 'CAT_FACT' as ConfigNodeName,
-    type: 'call',
-    url: 'https://catfact.ninja/fact',
-    input: 'request.body' as NameConnection,
-  },
-  {
-    name: 'DOG_FACT' as ConfigNodeName,
-    type: 'call',
-    url: 'https://dogapi.dog/api/v1/facts',
-  },
-  {
-    name: 'JOIN' as ConfigNodeName,
-    type: 'jq',
-    inputs: {
-      cat: 'CAT_FACT.body',
-      dog: 'DOG_FACT.body',
-    } as Record<FieldName, NameConnection>,
-    jq: '{\n  cat_fact: .cat.fact,\n  dog_fact: .dog.facts[0],\n}\n',
-  },
-  {
-    name: 'EXIT' as ConfigNodeName,
-    type: 'exit',
-    inputs: {
-      body: 'JOIN',
-    } as Record<FieldName, NameConnection>,
-    status: 200,
-  },
-  {
-    name: 'DANGLING_CALL' as ConfigNodeName,
-    type: 'call',
-    url: 'https://dogapi.dog/api/v1/facts',
-  },
-  {
-    name: 'READ_UPSTREAM' as ConfigNodeName,
-    type: 'jq',
-    input: 'service_response',
-    jq: '.',
-  },
-  {
-    name: 'VALUE' as ConfigNodeName,
-    type: 'static',
-    values: {
-      foo: 'bar',
-      baz: null,
-      nested: {
-        key: 'value',
-        array: [1, 2, 3],
-      },
-    },
-    outputs: {
-      foo: 'MY_CALL.body',
-      baz: 'MY_CALL.headers',
-    } as Record<FieldName, NameConnection>,
-  },
-  {
-    name: 'MY_CALL' as ConfigNodeName,
-    type: 'call',
-    url: 'https://example.com/api',
-  },
-], [
-  {
-    name: 'VALUE' as NodeName,
-    phase: 'response',
-    position: { x: 0, y: 0 },
-    fields: {
-      output: ['foo', 'baz', 'nested', 'unknown'] as FieldName[],
-    },
-    expanded: {
-      output: true,
-    },
-  },
-])
+function onChange(newConfig: DatakitConfig) {
+  emit('change', {
+    ...config,
+    ...newConfig,
+  })
+}
+
+const { modalOpen, load } = provideEditorStore(config?.nodes ?? [], [], { onChange })
+
+watch(() => config?.nodes, (newNodes) => {
+  // Only load if the modal is not open
+  if (modalOpen.value) {
+    return
+  }
+  load(newNodes ?? [], [])
+})
 </script>
+
+<style lang="scss" scoped>
+.dk-flow-editor {
+  .field {
+    margin-top: $kui-space-80;
+  }
+}
+</style>