+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.15.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.15.8...@kong-ui-public/entities-routes@3.15.9) (2025-08-18)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 ## [3.15.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.15.7...@kong-ui-public/entities-routes@3.15.8) (2025-08-15)